+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed the default relay to communities.nos.social.
 - Added "write post" tooltip to newly created group. [#81](https://github.com/verse-pbc/issues/issues/81)
 - Fixed an issue where groups created externally would not show up. [#146](https://github.com/verse-pbc/issues/issues/146)
+- Fix posting note on new group bugs. [#100](https://github.com/verse-pbc/issues/issues/100)
 - Fixed an issue where the group name would not show up. [#161](https://github.com/verse-pbc/issues/issues/161)
 
 ### Known Issues